+++ Cineaste/Extension/UIColor+Cineaste.swift
@@ -8,59 +8,57 @@
 
 import UIKit
 
-// The default values are needed due to this bug: 41244137
-// (Force unwrapping fails in Interface Builder)
-
+// swiftlint:disable force_unwrapping
 extension UIColor {
     // MARK: NavBar
-    static let cineNavBar = UIColor(named: "navBar") ?? .lightGray
-    static let cineNavBarTint = UIColor(named: "navBarTint") ?? .lightGray
-    static let cineNavBarTitle = UIColor(named: "navBarTitle") ?? .lightGray
+    static let cineNavBar = UIColor(named: "navBar")!
+    static let cineNavBarTint = UIColor(named: "navBarTint")!
+    static let cineNavBarTitle = UIColor(named: "navBarTitle")!
 
     // MARK: TabBar
-    static let cineTabBar = UIColor(named: "tabBar") ?? .lightGray
-    static let cineTabBarSelected = UIColor(named: "tabBarSelected") ?? .lightGray
-    static let cineTabBarNormal = UIColor(named: "tabBarNormal") ?? .lightGray
+    static let cineTabBar = UIColor(named: "tabBar")!
+    static let cineTabBarSelected = UIColor(named: "tabBarSelected")!
+    static let cineTabBarNormal = UIColor(named: "tabBarNormal")!
 
     // MARK: SegmentedControl
-    static let cineSegmentedControlTint = UIColor(named: "segmentedControlTint") ?? .lightGray
+    static let cineSegmentedControlTint = UIColor(named: "segmentedControlTint")!
 
     // MARK: Search TextField
-    static let cineSearchBackground = UIColor(named: "searchBackground") ?? .lightGray
-    static let cineSearchTint = UIColor(named: "searchTint") ?? .lightGray
-    static let cineSearchInput = UIColor(named: "searchInput") ?? .lightGray
+    static let cineSearchBackground = UIColor(named: "searchBackground")!
+    static let cineSearchTint = UIColor(named: "searchTint")!
+    static let cineSearchInput = UIColor(named: "searchInput")!
 
     // MARK: ToolBar
-    static let cineToolBarBackground = UIColor(named: "toolBarBackground") ?? .lightGray
-    static let cineToolBarTint = UIColor(named: "toolBarTint") ?? .lightGray
+    static let cineToolBarBackground = UIColor(named: "toolBarBackground")!
+    static let cineToolBarTint = UIColor(named: "toolBarTint")!
 
     // MARK: Safari
-    static let cineSafariBackground = UIColor(named: "safariBackground") ?? .lightGray
-    static let cineSafariTint = UIColor(named: "safariTint") ?? .lightGray
+    static let cineSafariBackground = UIColor(named: "safariBackground")!
+    static let cineSafariTint = UIColor(named: "safariTint")!
 
     // MARK: SwipeAction
-    static let cineSwipeSeen = UIColor(named: "swipeSeen") ?? .lightGray
-    static let cineSwipeWatchlist = UIColor(named: "swipeWatchlist") ?? .lightGray
-    static let cineSwipeDelete = UIColor(named: "swipeDelete") ?? .lightGray
+    static let cineSwipeSeen = UIColor(named: "swipeSeen")!
+    static let cineSwipeWatchlist = UIColor(named: "swipeWatchlist")!
+    static let cineSwipeDelete = UIColor(named: "swipeDelete")!
 
     // MARK: More...
-    static let cineListBackground = UIColor(named: "listBackground") ?? .lightGray
-    static let cineSeparator = UIColor(named: "separator") ?? .lightGray
-    static let cineCellBackground = UIColor(named: "cellBackground") ?? .lightGray
-    static let cineContentBackground = UIColor(named: "cellBackground") ?? .lightGray
-    static let cineVoteCircle = UIColor(named: "voteCircle") ?? .lightGray
-    static let cineVoteRectangle = UIColor(named: "voteRectangle") ?? .lightGray
-    static let cineButton = UIColor(named: "button") ?? .lightGray
-    static let cineButtonDark = UIColor(named: "buttonDark") ?? .lightGray
-    static let cineTitle = UIColor(named: "title") ?? .lightGray
-    static let cineDescription = UIColor(named: "description") ?? .lightGray
-    static let cineLink = UIColor(named: "link") ?? .lightGray
-    static let cineLinkOnBlack = UIColor(named: "linkOnBlack") ?? .lightGray
-    static let cineImageTint = UIColor(named: "imageTint") ?? .lightGray
-    static let cineAlertTint = UIColor(named: "alertTint") ?? .lightGray
-    static let cineShadow = UIColor(named: "shadow") ?? .lightGray
-    static let cineFooter = UIColor(named: "footer") ?? .lightGray
-    static let cineEmptyListDescription = UIColor(named: "emptyListDescription") ?? .lightGray
-    static let cineUsageActive = UIColor(named: "usageActive") ?? .lightGray
-    static let cineUsageInactive = UIColor(named: "usageInactive") ?? .lightGray
+    static let cineListBackground = UIColor(named: "listBackground")!
+    static let cineSeparator = UIColor(named: "separator")!
+    static let cineCellBackground = UIColor(named: "cellBackground")!
+    static let cineContentBackground = UIColor(named: "cellBackground")!
+    static let cineVoteCircle = UIColor(named: "voteCircle")!
+    static let cineVoteRectangle = UIColor(named: "voteRectangle")!
+    static let cineButton = UIColor(named: "button")!
+    static let cineButtonDark = UIColor(named: "buttonDark")!
+    static let cineTitle = UIColor(named: "title")!
+    static let cineDescription = UIColor(named: "description")!
+    static let cineLink = UIColor(named: "link")!
+    static let cineLinkOnBlack = UIColor(named: "linkOnBlack")!
+    static let cineImageTint = UIColor(named: "imageTint")!
+    static let cineAlertTint = UIColor(named: "alertTint")!
+    static let cineShadow = UIColor(named: "shadow")!
+    static let cineFooter = UIColor(named: "footer")!
+    static let cineEmptyListDescription = UIColor(named: "emptyListDescription")!
+    static let cineUsageActive = UIColor(named: "usageActive")!
+    static let cineUsageInactive = UIColor(named: "usageInactive")!
 }