+++ Cineaste/ViewController/SearchMovies/SearchSection.swift
@@ -1,3 +1,11 @@
+//
+//  SearchSection.swift
+//  Cineaste App
+//
+//  Created by Felizia Bernutz on 19.01.20.
+//  Copyright © 2020 spacepandas.de. All rights reserved.
+//
+
 enum SearchSection: Int {
     case tokens = 0
     case movies = 1